requirements.txt
zeroconf>=0.132.0
cryptography>=42.0.0
Pillow>=10.0.0
pystray>=0.19.0
ttkbootstrap>=1.10.0
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"